Add ability to skip surface regrid
An option, DOSFC, is added that allows the surface/nst regridding
to be skipped. The default is "YES" to maintain so the same
functionality is maintained. This is added to improve the efficiency
of some GEFS scripts.

Additionally, a new file, chgres_atm.log is created in OUTDIR to
signify the atmosphere portion is complete. This is also to aid in
GEFS efficiency.